+++ src/js/controller/handleTodoDoubleClick.js
@@ -1,0 +1,4 @@
+export const handleTodoDoubleClick = (event) => {
+  const li = event.target.closest("li");
+  li.classList.add("editing");
+};
